% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islandepi-package.R
\docType{package}
\name{islandepi-package}
\alias{islandepi}
\alias{islandepi-package}
\title{Epidemiological methods for Island Health}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

\code{islandepi} provides validated helpers for common applied epidemiology tasks
and standardized access to pinned BC population and health-geography
sources. It implements calculation mechanics and data access, not
disclosure or analytic policy. Confirm thresholds, standard populations and
interval methods against the standard governing each release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://temuulene.github.io/islandepi/}
  \item \url{https://github.com/temuulene/islandepi}
  \item Report bugs at \url{https://github.com/temuulene/islandepi/issues}
}

}
\author{
\strong{Maintainer}: Temuulen Enebish \email{temuulen@gmail.com}

}
\keyword{internal}
